webant: moved auth code in its own file
