+++ webant/util.py
@@ -1,7 +1,5 @@
 import functools
-from flask import send_file, session
-from authbone import Authenticator, Authorizator
-from users.api import get_user, get_anonymous_user, NotFoundException
+from flask import send_file
 
 
 def memoize(obj):
@@ -128,50 +126,3 @@ def get_centered_pagination(current, total, visible=5):
                 current = current,
                 last=last,
                 next = current+1 if(current < total) else None)
-
-
-class AuthtFromSession(Authenticator):
-
-    USERID_KEY = 'user_id'
-
-    def login(self, userID):
-        session[self.USERID_KEY] = userID
-
-    def logout(self):
-        session.pop(self.USERID_KEY)
-
-    def is_logged_in(self):
-        return self.USERID_KEY in session
-
-    def auth_data_getter(self):
-        return session.get(self.USERID_KEY, None)
-
-    def authenticate(self, userID):
-        try:
-            return get_user(id=userID)
-        except NotFoundException:
-            return None
-
-    def bad_auth_data_callback(self):
-        self.identity_elaborator(get_anonymous_user())
-
-    def not_authenticated_callback(self):
-        self.identity_elaborator(get_anonymous_user())
-
-
-class AuthzFromSession(Authorizator):
-
-    def check_capability(self, identity, capability):
-        return identity.can(capability[0], capability[1])
-
-
-class TransparentAutht(AuthtFromSession):
-
-    def perform_authentication(self, *args, **kwargs):
-        pass
-
-
-class TransparentAuthz(AuthzFromSession):
-
-    def perform_authorization(self, *args, **kwargs):
-        pass
